Removed Settings wheel
