+++ app/src/main/java/edu/csuci/lazynotetaker/feature_note/presentation/MainActivity.kt
@@ -8,8 +8,10 @@ import android.os.Bundle
 import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.material.Colors
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
 import androidx.core.net.toUri
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
@@ -94,4 +96,13 @@ class MainActivity : ComponentActivity() {
             }
         }
     }
+}
+@Composable
+fun lazynotetakerTheme(curTheme: Colors, content: @Composable() () -> Unit) {
+    MaterialTheme(
+        colors = MainActivity.appTheme,
+        typography = Typography,
+        shapes = Shapes,
+        content = content
+    )
 }
