+++ app/src/main/java/edu/csuci/lazynotetaker/feature_note/presentation/settings/components/UISection.kt
@@ -40,10 +40,7 @@ fun UISection(
             DefaultRadioButton(
                 text = "Dark",
                 selected = MainActivity.appTheme == DarkColorPalette,
-                onSelect = { onThemeChange
-
-
-                    MainActivity.appTheme = DarkColorPalette }
+                onSelect = { MainActivity.appTheme = DarkColorPalette }
             )
             Spacer(modifier = Modifier.width(0.dp))
             DefaultRadioButton(
